requirements.txt
pydantic>=2.0
pydantic-settings>=2.0
PyYAML>=6.0
Jinja2>=3.0
pytest>=7.0
